chore: minor cleanups on univ2 strategy code
