+++ src/protocols/interfaces/uni-v2-pool-data.interface.ts
@@ -1,0 +1,8 @@
+export interface UniV2PoolData {
+  contract: string;
+  token0: string;
+  token1: string;
+  reserve0: number;
+  reserve1: number;
+  totalSupply: number;
+}
